[U] AppConfig增加CleanMode选项，和ResConfig一致

+++ Assets/LiteQuark/Editor/Builder/AppBuild/AppPrepareStep.cs
@@ -15,7 +15,11 @@ namespace LiteQuark.Editor
 
         private void CleanOutputFile(ProjectBuilder builder)
         {
-            FileUtil.DeleteFileOrDirectory(builder.GetAppOutputPath());
+            if (builder.AppConfig.CleanBuildMode)
+            {
+                FileUtil.DeleteFileOrDirectory(builder.GetAppOutputPath());
+                AssetDatabase.Refresh();
+            }
         }
 
         private void ApplySetting(ProjectBuilder builder)
